requirements.txt
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
pdfkit==0.6.1
imgkit==1.2.2
django-filter==2.4.0
# psycopg2-binary==2.8.6
# psycopg2==2.8.6
